New panel icon to correct 16px visual result
Originally Gnome Shell's icon design guideline required 16px raster grid aligned icons for panel. As the original icon of the extensions had failed on this requirement I made a new simplified one to proper 16px visual result. The new icon slightly more follow the original Docker icon in it's curves but different at some other places as a requirement for low resolution. Sadly as the original icon can not properly shown on 16px grid and it needs to simplify some of it's original details. Again sadly the original icon was never properly usable at such low resolution and neither there is official alternative version for it. My current try on it has less details, and some curves aligned more to show a proper visual result. If Docker as a brand has a problem with it before they remove it please make an official raster aligned 16px alternate icon for their service.